quotapool: move benchmarks to a new file
Just code movement.

Release note: None
